Unify benchmarks with dataset selection
